+++ server/server-node/.gitignore
@@ -1,2 +1,3 @@
 # dependencies
 /node_modules
+.env
