+++ server/server-node/package.json
@@ -15,8 +15,9 @@
     "@types/cors": "^2.8.7",
     "@types/express": "^4.17.8",
     "@types/jsonwebtoken": "^8.5.0",
-    "@types/node": "^14.11.5",
+    "@types/node": "^14.11.10",
     "bcrypt": "^5.0.0",
+    "body-parser": "^1.19.0",
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
     "jsonwebtoken": "^8.5.1",
