+++ server/server-node/src/.env
@@ -1,2 +1,3 @@
-ACCESS_TOKEN_SECRET = 98336a1588f0622fc46d6d7a4975cbf9251e3879af897b13366d44b25a05c95f7b6ff25db3c65989c82daa39224d5828843fad13c081ffdfdd551584d7aa551d
-REFRESH_TOKEN_SECRET = 190666f80f3dfbe9de82c8c9044e20999072268908a446c6181027295042b4144ef9e8291163db5722fb935dd17b3158989de5f84e9d31eec82948660f0194c8
+ACCESS_TOKEN_SECRET=98336a1588f0622fc46d6d7a4975cbf9251e3879af897b13366d44b25a05c95f7b6ff25db3c65989c82daa39224d5828843fad13c081ffdfdd551584d7aa551d
+REFRESH_TOKEN_SECRET=190666f80f3dfbe9de82c8c9044e20999072268908a446c6181027295042b4144ef9e8291163db5722fb935dd17b3158989de5f84e9d31eec82948660f0194c8
+PORT=3000
