+++ server/server-node/src/controller/controller.ts
@@ -1,0 +1,1 @@
+// TODO tu urobiť operacie s databazou
