+++ server/server-node/src/middleware/logger.ts
@@ -1,0 +1,5 @@
+import express, { NextFunction } from 'express'
+
+export const logger = (request: express.Request, response: express.Response, next: NextFunction) => {
+    console.log(`${request.method} ${request.path} ${response}`);
+}
