+++ server/server-node/src/model/User.ts
@@ -1,0 +1,5 @@
+export interface User {
+    username: string,
+    password: string,
+    role: string
+}
