+++ server/server-node/src/routes/adminAuthentication.ts
@@ -1,0 +1,16 @@
+import { NextFunction } from "express";
+import jsonwebtoken from 'jsonwebtoken'
+
+const jwt = jsonwebtoken;
+
+export const authentication = (req: any, res:any, next: NextFunction) => {
+    const authHeader = req.headers[`authorizations`];
+    const token = authHeader && authHeader.split(' ')[1];
+    if(token == null) return res.sendStatus(`401`);
+    console.log(`Auth Token ${token}`)
+    jwt.verify(token, process.env.ACCESS_TOKEN_SECRET ?? '', (err: any, user: any) => {
+        if(err) return res.sendStatus(403);
+        req.user = user;
+        next();
+    });
+}
