+++ server/server-node/src/routes/populateUsers.ts
@@ -1,0 +1,5 @@
+import { User } from "../model/User";
+
+export const addAdmin = (usersArr: User[], user: User) => {
+    usersArr.push(user)
+}
